+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Bump rustc minimum version to 1.36
+
 ### Fixed
 - Fix rustc 1.36 `as_ref` [regression](https://github.com/rust-lang/rust/issues/60958)