+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add `vcs` format to the `script` command.
+
+### Changed
+- `script`: `vcom-arg` or `vlog-arg` can be used with format `vsim` or `vcs`.
 
 ## 0.15.0 - 2020-02-13
 ### Added